refactor form and actions to component
